requirements.txt
asgiref==3.2.7
Django==3.0.7
django-cors-headers==3.3.0
django-environ==0.4.5
django-filter==2.3.0
djangorestframework==3.11.0
Pillow==7.1.2
psycopg2-binary==2.8.5
pytz==2020.1
tzlocal==2.1
